lil update

+++ machines/iso/configuration.nix
@@ -70,28 +70,33 @@ let
     sleep 1
     url="https://github.com/DADA30000/dotfiles"
     clear
-    if gum confirm --default=false "Поменять URL репозитория с файлами конфигурации? (скрипт запускает start.sh из репозитория, репозиторий должен быть публичным)"; then
-      url=$(gum input --placeholder "Пример: https://github.com/DADA30000/dotfiles")
-    fi
-    if GIT_ASKPASS=true git ls-remote "$url" > /dev/null 2>&1; then
-      clear
-      echo -e "\e[34mКлонирование репозитория...\e[0m"
-      rm -rf /mnt2
-      mkdir /mnt2
-      if git clone "$url" --depth 1 /mnt2/dotfiles; then
-        cd /mnt2/dotfiles
-        echo -e "\e[34mЗапуск start.sh...\e[0m"
-        exec ./start.sh
-      else
-        echo -e "\e[31mОшибка клонирования репозитория, перезапуск скрипта...\e[0m"
-        sleep 3
+    if gum confirm --default=false "Использовать встроенный в ISO репозиторий?"; then
+      cd /repo
+      exec ./start.sh
+    else
+      if gum confirm --default=false "Поменять URL репозитория с файлами конфигурации? (скрипт запускает start.sh из репозитория, репозиторий должен быть публичным)"; then
+        url=$(gum input --placeholder "Пример: https://github.com/DADA30000/dotfiles")
+      fi
+      if GIT_ASKPASS=true git ls-remote "$url" > /dev/null 2>&1; then
+        clear
+        echo -e "\e[34mКлонирование репозитория...\e[0m"
         rm -rf /mnt2
+        mkdir /mnt2
+        if git clone "$url" --depth 1 /mnt2/dotfiles; then
+          cd /mnt2/dotfiles
+          echo -e "\e[34mЗапуск start.sh...\e[0m"
+          exec ./start.sh
+        else
+          echo -e "\e[31mОшибка клонирования репозитория, перезапуск скрипта...\e[0m"
+          sleep 3
+          rm -rf /mnt2
+          exec nix-install
+        fi
+      else
+        echo -e "\e[31mURL репозитория неверный или приватный, перезапуск скрипта...\e[0m"
+        sleep 3
         exec nix-install
       fi
-    else
-      echo -e "\e[31mURL репозитория неверный или приватный, перезапуск скрипта...\e[0m"
-      sleep 3
-      exec nix-install
     fi
   '';
   install-offline =
